+++ scripts/uninstall.js
@@ -5,48 +5,60 @@
 var fs = require('fs-extra');
 var path = require('path');
 var xml2js = require('xml2js');
-var cwd = process.cwd(); // $(project)/node_modules/cordova-uglify
+var utils = require('./utils');
 
-var uglifyJsPath = path.join(
-  cwd,
-  '../../',
-  'hooks',
-  'after_prepare',
-  'uglify.js'
-);
-var configFilePath = path.join(cwd, '../../', 'hooks', 'uglify-config.json');
+// __dirname = $(project)/node_modules/cordova-uglify/scripts
+var packageRoot = path.resolve(__dirname, '..');
+var projectRoot = utils.resolveProjectRoot(packageRoot);
 
-fs.unlinkSync(uglifyJsPath);
-fs.unlinkSync(configFilePath);
+var uglifyJsPath = path.join(projectRoot, 'hooks', 'after_prepare', 'uglify.js');
+var configFilePath = path.join(projectRoot, 'hooks', 'uglify-config.json');
 
-console.log('removed ' + uglifyJsPath + ' and ' + configFilePath);
-
-var cordovaConfigFilePath = path.join(cwd, '../../', 'config.xml'); // top-level config.xml
-var cordovaConfigFileData = fs.readFileSync(cordovaConfigFilePath);
-
-if (cordovaConfigFileData.indexOf('hooks/after_prepare/uglify.js') === -1) {
-  return;
+if (fs.existsSync(uglifyJsPath)) {
+  fs.unlinkSync(uglifyJsPath);
+  console.log('removed ' + uglifyJsPath);
 }
 
-var parser = new xml2js.Parser();
-parser.parseString(cordovaConfigFileData, function(err, result) {
-  if (err) {
-    console.log(err);
+if (fs.existsSync(configFilePath)) {
+  fs.unlinkSync(configFilePath);
+  console.log('removed ' + configFilePath);
+}
+
+var cordovaConfigFilePath = path.join(projectRoot, 'config.xml');
+
+function updateCordovaConfig() {
+  if (!fs.existsSync(cordovaConfigFilePath)) {
     return;
   }
 
-  var indexToDelete;
-  result.widget.hook.forEach(function(node, i) {
-    if (node.$.src === 'hooks/after_prepare/uglify.js') {
-      indexToDelete = i;
+  var cordovaConfigFileData = fs.readFileSync(cordovaConfigFilePath);
+
+  if (cordovaConfigFileData.indexOf('hooks/after_prepare/uglify.js') === -1) {
+    return;
+  }
+
+  var parser = new xml2js.Parser();
+  parser.parseString(cordovaConfigFileData, function(err, result) {
+    if (err) {
+      console.log(err);
+      return;
+    }
+
+    // Normalize hook to an array (xml2js may produce an object for a single entry)
+    var hooks = utils.normalizeHooks(result.widget.hook);
+
+    var indexToDelete = hooks.findIndex(function(node) {
+      return node && node.$ && node.$.src === 'hooks/after_prepare/uglify.js';
+    });
+
+    if (indexToDelete !== -1) {
+      hooks.splice(indexToDelete, 1);
+      result.widget.hook = hooks;
+      var builder = new xml2js.Builder();
+      var xml = builder.buildObject(result);
+      fs.writeFileSync(cordovaConfigFilePath, xml);
     }
   });
+}
 
-  if (indexToDelete > -1) {
-    result.widget.hook.splice(indexToDelete, 1);
-    var builder = new xml2js.Builder();
-    var xml = builder.buildObject(result);
-
-    fs.writeFileSync(cordovaConfigFilePath, xml);
-  }
-});
+updateCordovaConfig();
